Add Menu from Server

+++ AndroidFuelUp/Models/MainInfoAzs.cs
@@ -1,25 +1,12 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-using Android.App;
-using Android.Content;
-using Android.OS;
-using Android.Runtime;
-using Android.Views;
-using Android.Widget;
-
 namespace AndroidFuelUp.Models
 {
     public class MainInfoAzs
     {
         public string name { set; get; }
         public Ñoordinates coordinates { set; get; }
-        public int? codFuels { set; get; }
-        public int? codServices { set; get; }
+        public long? codFuels { set; get; }
+        public long? codServices { set; get; }
         public string operatorName { set; get; }
-
     }
 
     public class Ñoordinates
